embedded_systems_CV_project-main/Station6_AI/requirements.txt
opencv-python>=4.9.0
mediapipe>=0.10.14
numpy>=1.26.4,<2.0
pyserial>=3.5
pyttsx3>=2.90
pygame>=2.5.0

